Changed artifact id

+++ src/Company.SharedKernel/Handlers/DeleteAggregateRootCommand.cs
@@ -1,4 +1,4 @@
 ﻿namespace Company.SharedKernel.Abstractions;
 
-public record DeleteAggregateRootCommand<TAggregateRoot>(long Id) : IRequest<Result<bool>>
+public record DeleteAggregateRootCommand<TAggregateRoot>(int Id) : IRequest<Result<bool>>
     where TAggregateRoot : class, IAggregateRoot;
