+++ src/Company.Videomatic.Application/Features/Artifacts/Commands/DeleteArtifact.cs
@@ -1,6 +1,6 @@
 ﻿namespace Company.Videomatic.Application.Features.Artifacts.Commands;
 
-public record DeleteArtifactCommand(long Id) : DeleteAggregateRootCommand<Artifact>(Id);
+public record DeleteArtifactCommand(int Id) : DeleteAggregateRootCommand<Artifact>(Id);
 
 public class DeleteArtifactCommandValidator : AbstractValidator<DeleteArtifactCommand>
 {
