+++ src/Company.Videomatic.Application/Features/Playlists/Commands/DeletePlaylist.cs
@@ -1,6 +1,6 @@
 ﻿namespace Company.Videomatic.Application.Features.Playlists.Commands;
 
-public record DeletePlaylistCommand(long Id) : DeleteAggregateRootCommand<Playlist>(Id);
+public record DeletePlaylistCommand(int Id) : DeleteAggregateRootCommand<Playlist>(Id);
 
 internal class DeletePlaylistCommandValidator : AbstractValidator<DeletePlaylistCommand>
 {
